west: add flash-payload extension
Extending Zephyr's build system to "Flash" a payload generated by the
Hart Software Services' Payload Generator on to a target board. It
interacts with the Hart Software Services running on the target in a
"send" and "expect" manor, implemented with pyserial.

It guides the user through the process, prompting the user to
ensure that the device to write to is correct. Also, and very
importantly, since elevated privileges are required to write to a
mounted block device: it prompts the user to enter their password. It
hands passwords using getpass and encoding. It runs a subprocess with
these privileges to write to the eMMC.

Signed-off-by: Conor Paxton <conor.paxton@microchip.com>